requirements.txt
flask==2.3.3
langchain==0.0.353
langchain-openai==0.0.5
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
scikit-learn==1.3.0
textblob==0.17.1
vaderSentiment==3.3.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
transformers==4.35.0
torch==2.1.0
nltk==3.8.1
wordcloud==1.9.2
flask-cors==4.0.0
azure-ai-textanalytics==5.3.0
